Former-commit-id: e61a9775881d6664c17042234a2b5d275650897d

+++ src/DeltaShell.Plugins.NetworkEditor.Gui/Forms/ChartEditors/ShapeDeletePointTool.cs
@@ -24,14 +24,11 @@ namespace DeltaShell.Plugins.NetworkEditor.Gui.Forms.ChartEditors
                     if (null != ShapeModifyTool.ShapeFeatureEditor)
                     {
                         IPoint tracker = ShapeModifyTool.ShapeSelectTool.GetTrackerAt(ShapeModifyTool.ShapeFeatureEditor, tmP);
-                        if (null != tracker)
+                        if (null != tracker && ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
                         {
-                            if (ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
-                            {
-                                ShapeModifyTool.ShapeFeatureEditor.DeleteTracker(tracker);
-                                ((ShapeFeatureEditor) ShapeModifyTool.ShapeFeatureEditor).ShapeFeature.Invalidate();
-                                return;
-                            }
+                            ShapeModifyTool.ShapeFeatureEditor.DeleteTracker(tracker);
+                            ((ShapeFeatureEditor) ShapeModifyTool.ShapeFeatureEditor).ShapeFeature.Invalidate();
+                            return;
                         }
                     }
 
@@ -42,13 +39,10 @@ namespace DeltaShell.Plugins.NetworkEditor.Gui.Forms.ChartEditors
                     if (null != ShapeModifyTool.ShapeFeatureEditor)
                     {
                         IPoint tracker = ShapeModifyTool.ShapeSelectTool.GetTrackerAt(ShapeModifyTool.ShapeFeatureEditor, tmP);
-                        if (null != tracker)
+                        if (null != tracker && ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
                         {
-                            if (ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
-                            {
-                                c = ShapeModifyTool.ShapeFeatureEditor.GetCursor(tracker);
-                                return;
-                            }
+                            c = ShapeModifyTool.ShapeFeatureEditor.GetCursor(tracker);
+                            return;
                         }
 
                         c = Cursors.Default;
